Move OverlayType from design-system to elements package
